WebSocket Request Handler: Message Sending
Two-way communication between client and server is now functional. The code is in a bit of a mess as I was working hard to get websocket frame building working properly. Learning that the server should not mask when sending to the client is a big win. That simplified things a lot. Much cleanup to do around how the external server and server work together with the clients. Need to separate this out into a library. Great progress though. A really good night.